refactor! Hashable now returns `Result<Bytes>`
We also got rid of the `with_capacity` optimization, which was
implemented incorrectly. If I wanted to keep going down that path
for now I would have to use `prost::Message.encoded_len()`. In the
future if I need to revisit this I will probably pool the allocations
rather than allocate in the body of the function.